Refactor codebase into separate files, and enable `importsNotUsedAsValues: error`
